Rename dataclasses.py to schemas.py and rename models directory to recommender
